agregué archivo userCards.jss.ts para modularizar los estilos del organismo UserCards.tsx
