+++ src/organisms/userCards/userCards.jss.ts
@@ -1,0 +1,15 @@
+import { StyleSheet } from "react-native";
+ 
+const styles = StyleSheet.create({
+  grid: {
+    alignItems: "stretch",
+    justifyContent: "center",
+    flexWrap: "wrap",
+    flexDirection: "row",
+  },
+  item: {
+    margin: 8,
+  },
+});
+
+export default styles;
